procurement-app/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
supabase==2.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
numpy==1.26.0
scipy==1.11.0
PyJWT==2.8.0
reportlab==4.0.7
httpx>=0.24.0,<0.25.0
cachetools==5.3.2
aiosmtplib==3.0.1
jinja2==3.1.2
kaleido==0.2.1
